requirements.txt
# --- Core backend ---
Flask==3.0.3
Werkzeug==3.1.3
Flask-Cors==4.0.0
Flask-Limiter==3.7.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
gunicorn==21.2.0

# --- Database & ORM ---
SQLAlchemy==2.0.43
greenlet==3.2.4
psycopg2-binary==2.9.9
psycopg[binary]>=3.1

# --- Auth, Security & Password Hashing ---
passlib==1.7.4
PyJWT==2.9.0

# --- Utilities & Environment ---
python-dotenv==1.0.1
requests==2.32.5
typing_extensions==4.12.2
packaging==24.1

# --- CLI & Output Formatting ---
rich==13.9.4
Pygments==2.19.2
click==8.1.8

# --- Speedtest (optional, keep if using network speed checks) ---
speedtest-cli==2.1.3

# --- Video downloader (for future reels/shorts fetch bot) ---
yt-dlp==2025.9.5

# --- Background file watching / auto-reload (optional, dev use) ---
watchgod==0.8.2
redis>=4.5.0

# --- HTTP & Parsing ---
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.2.3

# --- Miscellaneous ---
blinker==1.9.0
itsdangerous==2.2.0
Markdown==3.7
markdown-it-py==4.0.0
mdurl==0.1.2
Deprecated==1.2.18
ordered-set==4.1.0
limits==5.6.0
setuptools==80.9.0
wheel==0.45.1
wrapt==1.17.3
